+++ js/src/main/scala/spatutorial/client/modules/MainMenu.scala
@@ -1,16 +1,20 @@
 package spatutorial.client.modules
 
+import japgolly.scalacss.ScalaCssReact._
 import japgolly.scalajs.react._
 import japgolly.scalajs.react.extra.OnUnmount
 import japgolly.scalajs.react.vdom.prefix_<^._
 import rx._
 import rx.ops._
+import spatutorial.client.components.Bootstrap.CommonStyle
 import spatutorial.client.components.Icon._
 import spatutorial.client.components._
 import spatutorial.client.services._
 import spatutorial.shared.TodoItem
 
 object MainMenu {
+  // shorthand for styles
+  @inline private def bss = GlobalStyles.bootstrapStyles
 
   case class Props(activeLocation: MainRouter.Loc, todos: Rx[Seq[TodoItem]])
 
@@ -19,7 +23,7 @@ object MainMenu {
   class Backend(t: BackendScope[Props, _]) extends OnUnmount {
     def mounted(): Unit = {
       // hook up to Todo changes
-      val obsItems = t.props.todos.foreach { _ => t.forceUpdate()}
+      val obsItems = t.props.todos.foreach { _ => t.forceUpdate() }
       onUnmount {
         // stop observing when unmounted (= never in this SPA)
         obsItems.kill()
@@ -33,7 +37,7 @@ object MainMenu {
     val todoCount = props.todos().count(!_.completed)
     Seq(
       <.span("Todo "),
-      if (todoCount > 0) <.span(^.className := "label label-danger label-as-badge", todoCount) else <.span()
+      if (todoCount > 0) <.span(bss.labelOpt(CommonStyle.danger), bss.labelAsBadge, todoCount) else <.span()
     )
   }
 
@@ -46,7 +50,7 @@ object MainMenu {
     .stateless
     .backend(new Backend(_))
     .render((P, _, B) => {
-    <.ul(^.className := "nav navbar-nav")(
+    <.ul(bss.navbar)(
       // build a list of menu items
       for (item <- menuItems) yield {
         <.li((P.activeLocation == item.location) ?= (^.className := "active"),