+++ js/src/main/scala/spatutorial/client/modules/MainRouter.scala
@@ -36,6 +36,7 @@ object MainRouter extends RoutingRules {
    */
   override protected def interceptRender(ic: InterceptionR) = {
     <.div(
+      // here we use plain Bootstrap class names as these are specific to the top level layout defined here
       <.nav(^.className := "navbar navbar-inverse navbar-fixed-top")(
         <.div(^.className := "container")(
           <.div(^.className := "navbar-header")(<.span(^.className := "navbar-brand")("SPA Tutorial")),